Add files via upload
This is a code that randomly simulates the winner of a tournament after a user inputs eight different clubs of their choice. If "re-shuffle" is selected; the draws are re-done, if "restart" is selected; the user starts over, and, if "yes" is selected; the tournament continues. It also has a loop that asks if the user wants to simulate again. And, ohh yes, it features a tournament history.