Add files via upload
This is a password generator that just like the name implies, helps the user to generate a password based on the user's preferences. Whether or not a digit, a symbol or an uppercase is in the password is entirely left to the user. With this, you definitely don't need to create a password yourself again.