Add files via upload
This is the password generator is the PYQT5 GUI fomat. It is even better than the Tkinter version of the password generator. It shows real time passwords once the button is pressed, allows for copying to clipboard, shows password strength in real time as each is generated. Try it out to see what I'm talking about!